refactor : 피드백 반영
- 예외 상황에 따른 예외 핸들링 세분화
- rest controller 별 중복 핸들링 url 클래스 위에 request mapping 애노테이션으로 분리
- page controller 핸들러중, 불필요 model 인자 제거
- path controller 최단 경로 api 핸들러 인자 유효성 검증 실패시 커스텀 예외 발생
- line station : 최단 경로 찾기 구현중 필요한 첫번째 구간 값 판별 기능 추가
- path search type enum : 최단 경로 검색 타입 별 그래프 구현시 가중치값 반환 로직 추가
- station : 역 목록중, 역의 id, 이름을 통해 역 검색 로직 위한 기능 추가
- stations : id, 이름을 통한 역 검색 기능을 위한 일급 컬렉션 추가
- subway edge : 그래프를 통한 최단 경로 검색시, edge로 활용할 클래스 추가
- error response : 예외 발생시 서버에서 예외에 대한 정보를 반환하기 위한 response 클래스 추가
- path request : 해당 dto 내 path search type 을 문자열 타입으로 변환
- 기타 custom exception : 기타 상황별 세부 커스텀 예외 추가
- station service : 역 생성시 location 헤더에 생성 역 자원 접근 위한 역 단일 조회 기능 추가
- path service : 알고리즘에 따른 유연한 경로 반환을 위한 인터페이스 추상화
- jgraph path service : path service 구현체로 jgraph 활용한 경로 찾는 서비스 클래스 추가
- js 파일들 : 서버에서 예외 발생시 클라이언트에서 에러 메시지 alert 시킬수 있도록 수정

+++ src/main/java/wooteco/subway/admin/service/StationService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.dto.StationResponse;
+import wooteco.subway.admin.exception.NotFoundResourceException;
 import wooteco.subway.admin.repository.StationRepository;
 
 @Transactional
@@ -18,8 +20,15 @@ public class StationService {
 	}
 
 	@Transactional(readOnly = true)
-	public List<Station> findAll() {
-		return stationRepository.findAll();
+	public List<StationResponse> findAll() {
+		return StationResponse.listOf(stationRepository.findAll());
+	}
+
+	@Transactional(readOnly = true)
+	public StationResponse findById(Long id) {
+		Station station = stationRepository.findById(id)
+			.orElseThrow(() -> new NotFoundResourceException("해당하는 역이 존재하지 않습니다."));
+		return StationResponse.of(station);
 	}
 
 	public Station save(Station station) {
@@ -27,6 +36,9 @@ public class StationService {
 	}
 
 	public void deleteById(Long id) {
+		if (!stationRepository.existsById(id)) {
+			throw new NotFoundResourceException("해당하는 역이 존재하지 않습니다.");
+		}
 		stationRepository.deleteById(id);
 	}
 }
